refactor : Car 클래스 테스트 코드 분리
